requirements.txt
# Core dependencies
python-dotenv==1.0.0
redis==5.0.1
websockets==12.0

# Blockchain & Web3
web3==6.11.3
eth-account==0.10.0

# API & HTTP
requests==2.31.0
aiohttp==3.9.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Logging & Monitoring
colorlog==6.8.0
python-telegram-bot==20.7

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1